pyproject.toml
[project]
name = "nonebot_plugin_suggarchat"
version = "0.1.5.6"
description = "Plugin for the Suggar chat framework compatible with Nonebot2."
authors = [{ name = "JohnRichard4096", email = "windowserror@163.com" }]
dependencies = [
    "nonebot2>=2.2.0",
    "openai>=1.5.0",
    "pydantic>=2.4.2",
    "nonebot-adapter-onebot>=2.0.0",
    "nonebot-plugin-localstore>=0.7",
]
requires-python = "~=3.9"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
keywords = ["chat", "suggar", "nonebot"]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[project.urls]
"Homepage" = "https://github.com/JohnRichard4096/nonebot_plugin_suggarchat"
"Source" = "https://github.com/JohnRichard4096/nonebot_plugin_suggarchat"
"Issue Tracker" = "https://github.com/JohnRichard4096/nonebot_plugin_suggarchat/issues"
